pyproject.toml
[tool.poetry]
name = "snowflakeconnector"
version = "0.1.0"
description = "Snowflake connector utilities for interfacing with Snowflake databases"
authors = ["sudheer-akkio <sudheer@akkio.com>"]
license = "Akkio"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
snowflake-connector-python = "^3.1.0"
numpy = "^1.25.2"
scikit-learn = "^1.3.0"
requests = "^2.31.0"
flask-wtf = "^1.1.1"
wtforms = "^3.0.1"
flask-bcrypt = "^1.0.1"
flask-login = "^0.6.2"
flask-sqlalchemy = "^3.0.5"
snowflake-sqlalchemy = "^1.5.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
snowflake-connector-python = { version = "^3.1.0", extras = ["pandas"] }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
